Add the python and icedoc artifacts to the omeroBlitz publication
In terms of deployment, this means all artifacts (Java, Python, Icedoc)
should be published in a single transaction. For SNAPSHOT artifacts,
this should now avoid the creation of three successive iterations. For
release artifacts, this should remove the requirement for a privileged
user as the POM should not be overwritten.