Explicitly include details.externalInfo when relinking duplicate objects
The current implementation will duplicate objects with ExternalInfo objects
associated to them and create both the original object as well as a duplicated
ExternalInfo object. However, the linking between the duplicates is missing since
the externalInfo link is stored under `details.externalInfo`.

This commit should allow to properly handle this scenario and relink the duplicated
objects.